Discovered why encrypted preferences weren't being stored in the contract
AES.encrypt returns a CipherParams object which the contract is unable to store, hence why retrieval was impossible. Storing just the ciphertext however doesn't work as AES.decrypt also requires a CipherParams object. So either we need to work out a different way to store encrypted preferences to enable the storage of CipherParams or a different encryption module needs to be chosen.